requirements.txt
anthropic
streamlit
requests
python-dotenv